% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govscienceuseR-package.R
\docType{package}
\name{govscienceuseR-package}
\alias{govscienceuseR-package}
\alias{govscienceuseR}
\title{Easily Install and Load the \code{govscienceuseR} Tools for automated extraction
and disambiguation of scientific resources cited in government documents.}
\description{
\code{govscienceuseR} is a collection of software packages for bibliometric
analysis of policy documents that are designed to work together.
This package is designed to make it easy to install
and load multiple \code{govscienceuseR} packages in a single step.
}
\details{
\code{govscienceuseR} software implements advances in tagging, classifying,
and disambiguating references found in unstructured text (e.g., PDFs).
Together, the set of packages provide a workflow that ultimately allows
the user to perform a range of bibliometric analyses.

The code is actively developed and maintained by the \code{govscienceuseR}
development team. New functionality is being added over time.

\code{govscienceuseR} packages are primarily written in \code{R}.
The govscienceuseR project uses an
open development process for the packages, hosted
on GitHub \url{https://github.com/govscienceuseR}, and contributions can be made
via pull requests.

Training materials are also available online,
please see the \code{govscienceuseR} project
website at \url{https://www.govscienceuseR.github.io} for more information.

The full suite of packages has the following components (those automatically
downloaded with the \pkg{govscienceuseR} package are noted):

\itemize{

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\item \pkg{referenceExtract} 
Process PDFs and tag citations/references observed in PDFs.
(automatically downloaded)

\\item \pkg{referenceClassify}
Clean and classify citations by category (e.g., academic journal, agency document)  
(automatically downloaded)

\\item \pkg{indexBuild}
Create a database of academic work to search against for disambiguating extracted citations. 
(automatically downloaded)

\\item \pkg{referenceSearch}
Search extracted citations against indexed database of canonical citations to match and disambiguate extracted citations.
(automatically downloaded)
}\if{html}{\out{</div>}}

}
\code{citation("packagename")}.  Thank you!
}
\author{
govscienceuseR Development Team \email{contact@govscienceuseR.org}

Maintainer: Tyler A. Scott \email{tascott@ucdavis.edu}
}
