% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_govscienceuseR.R
\name{install_govscienceuseR}
\alias{install_govscienceuseR}
\title{Install the \pkg{govscienceuseR} toolkit}
\usage{
install_govscienceuseR(..., ask = FALSE, checkBuilt = TRUE)
}
\arguments{
\item{\dots}{Additional arguments to be passed to
\code{\link[devtools-remote-reexports]{devtools::install_github}}.}

\item{ask, checkBuilt}{Arguments to \code{\link[devtools-remote-reexports]{devtools::update_packages}}
documentation. The defaults are different from those of that function.}
}
\value{
\code{install_govscienceuseR} returns NULL invisibly.
}
\description{
A wrapper around \code{\link[devtools]{install_github}} to install the core component
packages of the \pkg{govscienceuseR}.
}
\details{
Installes the core component packages of govscienceuseR toolkit, using
\code{\link[utils]{setRepositories}} and \code{\link[devtools]{update_packages}}.  For
the list of packages automatically updated, see \code{\link[govscienceuseR]{govscienceuseR}}.

With no additional arguments specified, this will install from the binaries hosted at
\url{https://govscienceuseR.github.io}. Please report any bugs on
the GitHub package repository.
}
\examples{

\dontrun{
# Install from Github
govscienceuseR::install_govscienceuseR()
}

}
\seealso{
\code{\link[devtools-remote-reexports]{devtools::install_github}}, \code{\link{setRepositories}}
}
\keyword{utilities}
