% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_govscienceuseR.R
\name{update_govscienceuseR}
\alias{update_govscienceuseR}
\title{Update the \pkg{govscienceuseR} toolkit}
\usage{
update_govscienceuseR(..., ask = FALSE, checkBuilt = TRUE)
}
\arguments{
\item{\dots}{Additional arguments to be passed to
\code{\link{update_packages}}.}

\item{ask, checkBuilt}{Arguments to \code{\link{devtools::update_packages}}
documentation. The defaults are different from those of that function.}
}
\value{
\code{update_govscienceuseR} returns NULL invisibly.
}
\description{
A wrapper around \code{\link{devtools::update_packages()}} to update the core component
packages of the \pkg{govscienceuseR} to their latest versions.
}
\details{
Updates the core component packages of govscienceuseR toolkit, using
\code{\link[utils]{setRepositories}} and \code{\link[devtools]{update_packages}}.  For
the list of packages automatically updated, see \code{\link[govscienceuseR]{govscienceuseR}}.

This function should be called immediately after restarting R,
since there are no good ways to update packages once they are loaded.

With no additional arguments specified, the function will update the
packages from Github. This will install from the binaries hosted at
\url{https://govscienceuseR.github.io}. Please report any bugs on
the GitHub package repository.
}
\examples{

\dontrun{
# Update from Github
govscienceuseR::update_govscienceuseR()
}

}
\seealso{
\code{\link{setRepositories}}, \code{\link{update_packages}},
\code{\link{install.packages}}
}
\keyword{utilities}
